Modify:컴포넌트분리완료
